Move the keep_alive scheduler into serve/sched.rs
